cleanup: cleaning up the tests
