+++ test/tests/e2e-workflow.ts
@@ -6,6 +6,7 @@ import * as os from 'os'
 import {ABI, APIClient, FetchProvider, Serializer} from '@wharfkit/antelope'
 import fetch from 'node-fetch'
 import {log} from '../../src/utils'
+import {killProcessAtPort} from '../utils/test-helpers'
 
 /**
  * E2E tests for the complete workflow:
@@ -52,121 +53,17 @@ suite('E2E Workflow', () => {
         originalHome = process.env.HOME || ''
         process.env.HOME = testDir
 
-        // Kill any existing processes on port 8888
+        // Stop any existing chain before starting
+        // Try chain local stop first (works if chain was started with same HOME)
         try {
-            // Try to stop cleanly first
             execSync(`node ${cliPath} chain local stop`, {encoding: 'utf8', stdio: 'ignore'})
-            // Give it a moment to fully shut down
             execSync('sleep 1', {encoding: 'utf8', stdio: 'ignore'})
-        } catch (error) {
-            // Ignore errors
+        } catch {
+            // Ignore errors - chain might not be running or was started with different HOME
         }
 
-        // Also check for PID file and kill that process directly
-        try {
-            const pidFile = path.join(os.homedir(), '.wharfkit', 'chain', 'nodeos.pid')
-            if (fs.existsSync(pidFile)) {
-                const pid = parseInt(fs.readFileSync(pidFile, 'utf-8').trim(), 10)
-                if (!isNaN(pid) && pid > 0) {
-                    try {
-                        execSync(`kill -9 ${pid}`, {encoding: 'utf8', stdio: 'ignore'})
-                        // Remove the PID file
-                        fs.unlinkSync(pidFile)
-                        // Give it a moment to fully shut down
-                        execSync('sleep 0.5', {encoding: 'utf8', stdio: 'ignore'})
-                    } catch {
-                        // Process might be gone already
-                    }
-                }
-            }
-        } catch (error) {
-            // Ignore errors
-        }
-
-        // Force kill only if it's nodeos (only check for LISTENING processes, not client connections)
-        try {
-            // Use -sTCP:LISTEN to only find processes LISTENING on the port, not clients
-            const pids = execSync(`lsof -ti:8888 -sTCP:LISTEN`, {encoding: 'utf8'})
-                .trim()
-                .split('\n')
-            for (const pid of pids) {
-                if (!pid) continue
-                const pidNum = parseInt(pid)
-                if (isNaN(pidNum)) continue
-
-                // Never kill our own process tree
-                if (pidNum === process.pid || pidNum === process.ppid) continue
-
-                try {
-                    // Check if process is nodeos
-                    const cmd = execSync(`ps -p ${pidNum} -o command=`, {encoding: 'utf8'}).trim()
-                    if (cmd.includes('nodeos')) {
-                        // Only kill nodeos processes
-                        execSync(`kill -9 ${pidNum}`, {encoding: 'utf8', stdio: 'ignore'})
-                    } else {
-                        // Log what we found but didn't kill
-                        // eslint-disable-next-line no-console
-                        console.log(
-                            `Found non-nodeos process ${pidNum} listening on port 8888: ${cmd}`
-                        )
-                    }
-                } catch {
-                    // Process might be gone already
-                }
-            }
-        } catch (error) {
-            // Ignore errors (lsof fails if no process found)
-        }
-
-        // Wait for port 8888 to be free (only check for LISTENING processes)
-        const startTime = Date.now()
-        while (Date.now() - startTime < 20000) {
-            // Increased wait to 20s
-            try {
-                execSync('lsof -ti:8888 -sTCP:LISTEN', {encoding: 'utf8', stdio: 'ignore'})
-                // Port is still in use, wait for it to be released
-                // If it's been more than 2 seconds, try stopping again
-                if (Date.now() - startTime > 2000) {
-                    try {
-                        execSync(`node ${cliPath} chain local stop`, {
-                            encoding: 'utf8',
-                            stdio: 'ignore',
-                        })
-                    } catch {
-                        // Ignore errors
-                    }
-                }
-                execSync('sleep 0.5')
-            } catch {
-                // lsof failed, meaning port is free
-                break
-            }
-        }
-
-        // Final check - verify port is free (only LISTENING processes)
-        try {
-            const remainingPids = execSync('lsof -ti:8888 -sTCP:LISTEN', {encoding: 'utf8'}).trim()
-            if (remainingPids) {
-                // Check what's still holding the port
-                const pids = remainingPids.split('\n')
-                const processes: string[] = []
-                for (const pid of pids) {
-                    if (!pid) continue
-                    try {
-                        const cmd = execSync(`ps -p ${pid} -o command=`, {encoding: 'utf8'}).trim()
-                        processes.push(`${pid}: ${cmd}`)
-                    } catch {
-                        processes.push(`${pid}: (unknown)`)
-                    }
-                }
-                throw new Error(
-                    `Port 8888 is still in use after cleanup by: ${processes.join(', ')}`
-                )
-            }
-        } catch (e: any) {
-            if (e.message && e.message.includes('Port 8888 is still in use')) throw e
-            // lsof failed, meaning port is free - this is what we want
-        }
+        // Also check port 8888 directly in case chain was started by another test with different HOME
+        killProcessAtPort(8888)
 
         execSync(`node ${cliPath} chain local start`, {encoding: 'utf8'})
     })
@@ -181,11 +78,7 @@ suite('E2E Workflow', () => {
             fs.rmSync(testDir, {recursive: true, force: true})
         }
 
-        try {
-            execSync(`node ${cliPath} chain local stop`, {encoding: 'utf8'})
-        } catch (e) {
-            // Ignore error
-        }
+        execSync(`node ${cliPath} chain local stop`, {encoding: 'utf8'})
     })
 
     suite('Wallet Key Management', () => {
